edited placement #52
