+++ routes/placements.js
@@ -11,6 +11,43 @@ router.route('/')
     }, function(err) {
       res.writeStatus(400);
     });
+    var adID = winningAd.id;
+    var price = winningAd.price;
+    client.search({
+    index: 'index',
+    type: 'creative',
+    body: {
+      query: {
+        match: {
+          id: adID
+        }
+      }
+    }
+  }).then(function (resp) {
+     var hits = resp.hits.hits;
+      var cID = hits.get(campaignId);
+      client.search({
+    index: 'index',
+    type: 'campaign',
+    body: {
+      query: {
+        match: {
+          id: cID
+        }
+      }
+    }
+  }).then(function (resp) {
+     var hits = resp.hits.hits;
+      var budget = hits.get(budget);
+      var newBudget = budget - price ;
+      
+  }, function (err) {
+      console.trace(err.message);
+  });
+  }, function (err) {
+      console.trace(err.message);
+  });
+  (winningAd).data('budget', "newBudget");
   });
 
 module.exports = router;
